+++ scripts/functions.sh
@@ -56,7 +56,7 @@ function rsync_files()
   git ls-files "${root_dir}/**/.rsync" | while read rsync_file; do
     pushd "$( dirname "${rsync_file}" )"
     cat .rsync | while read line; do
-      rsync -ah --delete ${line}
+      rsync --archive ${line}
     done
     popd
   done
@@ -126,7 +126,7 @@ function push_subrepo_branch_maybe()
   branch=$(git_branch)
   
   if is_branch_or_tag_for_subrepo "${branch}" "${subrepo}"; then
-    git push "${remote}" $(splitsh-lite --prefix=${subrepo}):refs/heads/${branch}
+    git push --force "${remote}" $(splitsh-lite --prefix=${subrepo}):refs/heads/${branch}
   fi
 }
 
@@ -137,7 +137,7 @@ function push_subrepo_tag_maybe()
   if [ -z "${TRAVIS_TAG}" ]; then
     echo "No tags to push"
   elif is_branch_or_tag_for_subrepo "${TRAVIS_TAG}" "${subrepo}"; then
-    git push "${remote}" $(splitsh-lite --prefix=${subrepo} --origin=refs/tags/${TRAVIS_TAG}):refs/tags/${TRAVIS_TAG}
+    git push --force "${remote}" $(splitsh-lite --prefix=${subrepo} --origin=refs/tags/${TRAVIS_TAG}):refs/tags/${TRAVIS_TAG}
   fi
 }