+++ cucumber-expressions/CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    [aslakhellesoy])
 
 ### Fixed
+* The built-in `{word}` parameter type handles unicode (any non-space character)
+  ([#471](https://github.com/cucumber/cucumber/pull/471)
+   [savkk]
+   [aslakhellesoy]
 * Parenthesis inside character class should not be treated as capture group.
   ([#454](https://github.com/cucumber/cucumber/issues/454)
    [#461](https://github.com/cucumber/cucumber/pull/461)
@@ -485,4 +489,5 @@ N/A
 [luke-hill]:        https://github.com/luke-hill
 [mpkorstanje]:      https://github.com/mpkorstanje
 [kAworu]:           https://github.com/kAworu
+[savkk]:            https://github.com/savkk
 [spicalous]:        https://github.com/spicalous